% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-advanced-batch.R
\name{advanced_batch_search}
\alias{advanced_batch_search}
\title{Advanced Batch Search using CEU Mass Mediator API}
\usage{
advanced_batch_search(
  cmm_url = "https://ceumass.eps.uspceu.es/mediator/api/v3/advancedbatch",
  chemical_alphabet = "all",
  modifiers_type = "none",
  metabolites_type = "all-except-peptides",
  databases = list("hmdb"),
  masses_mode = "mz",
  ion_mode = "positive",
  adducts = list("all"),
  deuterium = FALSE,
  tolerance = 10,
  tolerance_mode = "ppm",
  masses = c(399.3367, 421.3169, 315.2424, 337.2234, 280.2402),
  all_masses = list(),
  retention_times = c(18.8425, 18.8425, 8.1449, 8.1449, 28.2695, 4.0216),
  all_retention_times = list(),
  composite_spectra = list(),
  all_composite_spectra = list()
)
}
\arguments{
\item{cmm_url}{A character string specifying the CEU Mass Mediator API endpoint.}

\item{chemical_alphabet}{A character string specifying the chemical alphabet to use.}

\item{modifiers_type}{A character string specifying the modifier type.}

\item{metabolites_type}{A character string specifying the metabolites type.}

\item{databases}{A list specifying the databases to search.}

\item{masses_mode}{A character string specifying the masses mode.}

\item{ion_mode}{A character string specifying the ionization mode.}

\item{adducts}{A list specifying the adducts to include in the search.}

\item{deuterium}{Logical value indicating whether to consider deuterium substitutions.}

\item{tolerance}{A numeric value specifying the mass tolerance.}

\item{tolerance_mode}{A character string specifying the tolerance mode.}

\item{masses}{A numeric vector of masses to search.}

\item{all_masses}{A list for arrays of masses (optional).}

\item{retention_times}{A numeric vector of retention times corresponding to the masses.}

\item{all_retention_times}{A list for retention times arrays (optional).}

\item{composite_spectra}{A list for composite spectra (optional).}

\item{all_composite_spectra}{A list for all composite spectra (optional)}
}
\value{
A dataframe containing the search results from the CEU Mass Mediator API.
}
\description{
\code{advanced_batch_search} performs an advanced batch search on the CEU Mass Mediator API
and returns a dataframe of search results.
}
\examples{
\donttest{
df <- advanced_batch_search(
cmm_url = "https://ceumass.eps.uspceu.es/mediator/api/v3/advancedbatch",
chemical_alphabet = "all",
modifiers_type = "none",
metabolites_type = "all-except-peptides",
databases = list("hmdb"),
masses_mode = "mz",
ion_mode = "positive",
adducts = list("all"),
deuterium = FALSE,
tolerance = 10.0,
tolerance_mode = "ppm",
masses = c(399.3367, 421.31686, 315.2424, 337.2234, 280.2402),
all_masses = list(),
retention_times = c(18.842525, 18.842525, 8.144917, 8.144917, 28.269503, 4.021555),
all_retention_times = list(),
composite_spectra = list(),
all_composite_spectra = list()
)

head(df)

}


}
