% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-advanced-batch-body.R
\name{create_advanced_batch_body}
\alias{create_advanced_batch_body}
\title{Create POST request Body for advanced batch search}
\usage{
create_advanced_batch_body(
  chemical_alphabet = "all",
  modifiers_type = "none",
  metabolites_type = "all-except-peptides",
  databases = list("hmdb"),
  masses_mode = "mz",
  ion_mode = "positive",
  adducts = list("all"),
  deuterium = FALSE,
  tolerance = 10,
  tolerance_mode = "ppm",
  masses = c(399.3367, 421.3169, 315.2424, 337.2234, 280.2402),
  all_masses = list(),
  retention_times = c(18.8425, 18.8425, 8.1449, 8.1449, 28.2695, 4.0216),
  all_retention_times = list(),
  composite_spectra = list(),
  all_composite_spectra = list()
)
}
\arguments{
\item{chemical_alphabet}{A string for the chemical alphabet: "CHNOPS", "CHNOPSCL", "ALL".}

\item{modifiers_type}{A string for the modifier type: "none", "NH3", "HCOO", etc.}

\item{metabolites_type}{A string for metabolites type: "all-except-peptides", "only-lipids", etc.}

\item{databases}{A vector of databases to search: e.g., list("HMDB").}

\item{masses_mode}{A string specifying masses mode: "neutral" or "mz".}

\item{ion_mode}{A string for the ionization mode: "positive", "negative", etc.}

\item{adducts}{A vector of adducts to include in the search, e.g., list("M+H", "M+Na").}

\item{deuterium}{Logical value: whether to consider deuterium substitutions (TRUE or FALSE).}

\item{tolerance}{Numeric value specifying the tolerance.}

\item{tolerance_mode}{A string for the tolerance mode: "ppm" or "mDa".}

\item{masses}{Numeric vector of masses to search.}

\item{all_masses}{Empty array for all masses (optional).}

\item{retention_times}{Numeric vector of retention times.}

\item{all_retention_times}{Empty array for retention times (optional).}

\item{composite_spectra}{Empty array for composite spectra.}

\item{all_composite_spectra}{A list for all composite spectra (optional)}
}
\value{
A properly formatted JSON body for the POST request.
}
\description{
\code{create_advanced_batch_body} creates a list for the advanced batch search POST request.
}
