% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-msms-body.R
\name{create_msms_body}
\alias{create_msms_body}
\title{Create MS/MS search POST request body}
\usage{
create_msms_body(
  ion_mass,
  ms_ms_peaks,
  precursor_ion_tolerance = 500,
  precursor_ion_tolerance_mode = "mDa",
  precursor_mz_tolerance = 1000,
  precursor_mz_tolerance_mode = "mDa",
  ion_mode = "positive",
  ionization_voltage = "all",
  spectra_types = "experimental"
)
}
\arguments{
\item{ion_mass}{ion_mass}

\item{ms_ms_peaks}{ms_ms_peaks}

\item{precursor_ion_tolerance}{precursor_ion_tolerance}

\item{precursor_ion_tolerance_mode}{precursor_ion_tolerance_mode}

\item{precursor_mz_tolerance}{precursor_mz_tolerance}

\item{precursor_mz_tolerance_mode}{precursor_mz_tolerance_mode}

\item{ion_mode}{ion_mode}

\item{ionization_voltage}{ionization_voltage}

\item{spectra_types}{spectra_types}
}
\value{
If all inputs are all correctly formatted, a string of a POST request will be returned for the result.
}
\description{
\code{create_msms_body} returns a string of a POST request body.
}
