% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcm.R
\name{jmcm.kron}
\alias{jmcm.kron}
\title{Fit Kronecker Product based Covariance Structure Models}
\usage{
jmcm.kron(formula, data = NULL, bcov.method = c("mcd", "acd", "hpc"),
  wcov.method = c("mcd", "acd", "hpc"), control = jmcmControl(),
  start = NULL)
}
\arguments{
\item{data}{a data frame containing the variables named in formula.}

\item{bcov.method}{between-subject covariance structure modelling method,
choose 'mcd' (Pourahmadi 1999), 'acd' (Chen and Dunson 2013) or 'hpc'
(Zhang et al. 2015).}

\item{wcov.method}{within-subject covariance structure modelling method,
choose 'mcd' (Pourahmadi 1999), 'acd' (Chen and Dunson 2013) or 'hpc'
(Zhang et al. 2015).}

\item{start}{starting values for the parameters in the model.}
}
\description{
Fit a kronecker product based covariance structure model to
multivariate longitudinal data with multiple responses.
}
