% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmmrControl.R
\name{kcmmrControl}
\alias{kcmmrControl}
\title{Control of Kronecker Product based Covariance Structure Model Fitting}
\usage{
kcmmrControl(trace = FALSE)
}
\arguments{
\item{trace}{whether or not the value of the objective function and the
parameters should be print on every trace'th iteration.}
}
\description{
Construct control structures for kronecker product based
covariance structure model fitting
}
