% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmmr.R
\name{mcmmr}
\alias{mcmmr}
\title{Fit MCBD-based Covariance Matrices Models}
\usage{
mcmmr(formula, data = NULL, quad = c(3, 3, 3, 3), cov.method = c("mcd",
  "acd", "hpc"), control = mcmmrControl(), start = NULL)
}
\arguments{
\item{data}{a data frame containing the variables named in formula.}

\item{cov.method}{covariance structure modelling method for matrix Dt,
choose 'mcd' (Pourahmadi, 1999), 'acd' (Chen and Dunson, 2013) or 'hpc'
(Zhang et al. 2015).}

\item{control}{a list (of correct class, resulting from mcmmrControl())
containing control parameters, see the *mcmmrControl documentation for 
details.}

\item{start}{starting values for the parameters in the model.}
}
\description{
Fit a modified Cholesky block decomposition (MCBD) based 
 covariance matrices Models to longitudinal data with multiple responses.
}
