% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcm.R
\name{modular}
\alias{modular}
\alias{mldFormula}
\alias{optimizeMcbd}
\alias{mkMcbdMod}
\title{Modular Functions for Covariance Matrices Model Fits}
\usage{
mldFormula(formula, data = NULL, quad = c(3, 3, 3, 3),
  cov.method = c("mcd", "acd", "hpc"), control = jmcmControl(),
  start = NULL)

optimizeMcbd(m, Y, X, U, V, W, time, cov.method, control, start)

mkMcbdMod(opt, args, quad, cov.method, mc)
}
\arguments{
\item{formula}{}
}
\description{
Modular function for covariance matrices model fits
}
