% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcm.R
\docType{data}
\name{poplar}
\alias{poplar}
\title{Poplar Growth Data}
\format{A data frame with 660 rows and 4 variables}
\usage{
data(poplar)
}
\description{
The attribute data are measurements of stem height and stem diameter for 78 
poplar (genus Populus) trees over time. Measurements were taken at the end of
each year, for 11 years. Thus we have N = 78, T = 11, and J = 2. 
No covariates, apart from time itself, were measured
}
\details{
\itemize{
  \item id: subject id
  \item year: measurement time
  \item height: stem height for poplar tree
  \item diameter: stem diameter for poplar tree
}
}
\keyword{datasets}
