% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcm.R
\docType{methods}
\name{show,mcbdMod-method}
\alias{show,mcbdMod-method}
\title{Print information for mcbdMod-class}
\usage{
\S4method{show}{mcbdMod}(object)
}
\arguments{
\item{object}{a fitted joint mean covariance model of class "mcbdMod", i.e.,
typically the result of jmcm.mcbd().}
}
\description{
Print information for mcbdMod-class
}
